requirements.txt
fastapi 
uvicorn 
python-dotenv
langchain
langchain-community
langchain-core
langgraph
langchain-google-genai
faiss-cpu
PyPDF
pydantic
python-multipart
sentence-transformers
langchain_groq
langchain_huggingface